generalize burrow code to other unit assignment screens
